build: fix the trixie brought up warning->error issues

+++ modules/http/http-worker.c
@@ -283,7 +283,7 @@ _setup_static_options_in_curl(HTTPDestinationWorker *self)
   if (owner->accept_redirects)
     {
       curl_easy_setopt(self->curl, CURLOPT_FOLLOWLOCATION, 1L);
-      curl_easy_setopt(self->curl, CURLOPT_POSTREDIR, CURL_REDIR_POST_ALL);
+      curl_easy_setopt(self->curl, CURLOPT_POSTREDIR, (long) CURL_REDIR_POST_ALL);
 #if SYSLOG_NG_HAVE_DECL_CURLOPT_REDIR_PROTOCOLS_STR
       curl_easy_setopt(self->curl, CURLOPT_REDIR_PROTOCOLS_STR, "http,https");
 #else
